Crate antidote

Mutex and RwLock types that do not poison themselves.

These types expose identical APIs to the standard library Mutex and RwLock except that they do not return PoisonErrors.

Structs

Condvar

Like std::sync::Condvar.

Mutex

Like std::sync::Mutex except that it does not poison itself.

MutexGuard

Like std::sync::MutexGuard.

RwLock

Like std::sync::RwLock except that it does not poison itself.

RwLockReadGuard

Like std::sync::RwLockReadGuard.

RwLockWriteGuard

Like std::sync::RwLockWriteGuard.

TryLockError

Like std::sync::TryLockError.

WaitTimeoutResult

A type indicating whether a timed wait on a condition variable returned due to a time out or not.

Type Definitions

TryLockResult

Like std::sync::TryLockResult.